moved flow extensions to their related classes, refactor receiveCredential function, refactor other.kt
