+++ cordlib-contracts-states/src/main/kotlin/com/luxoft/blockchainlab/corda/hyperledger/indy/data/state/IndyCredential.kt
@@ -4,8 +4,14 @@ import com.luxoft.blockchainlab.corda.hyperledger.indy.data.schema.CredentialSch
 import com.luxoft.blockchainlab.hyperledger.indy.models.CredentialInfo
 import com.luxoft.blockchainlab.hyperledger.indy.models.CredentialRequestInfo
 import net.corda.core.contracts.LinearState
+import net.corda.core.contracts.StateAndRef
 import net.corda.core.contracts.UniqueIdentifier
+import net.corda.core.flows.FlowLogic
 import net.corda.core.identity.AbstractParty
+import net.corda.core.node.services.Vault
+import net.corda.core.node.services.queryBy
+import net.corda.core.node.services.vault.Builder.equal
+import net.corda.core.node.services.vault.QueryCriteria
 import net.corda.core.schemas.MappedSchema
 import net.corda.core.schemas.PersistentState
 import net.corda.core.schemas.QueryableState
@@ -37,4 +43,24 @@ open class IndyCredential(
     }
 
     override fun supportedSchemas(): Iterable<MappedSchema> = listOf(CredentialSchemaV1)
+}
+
+/**
+ * This method is used to get indy credential state from vault
+ *
+ * @param id                id of credential
+ *
+ * @return                  corda state of indy credential or null if none exists
+ */
+fun FlowLogic<Any>.getIndyCredentialState(id: String): StateAndRef<IndyCredential>? {
+    val generalCriteria =
+        QueryCriteria.VaultQueryCriteria(Vault.StateStatus.UNCONSUMED)
+    val existingId = QueryCriteria.VaultCustomQueryCriteria(
+        CredentialSchemaV1.PersistentCredential::id.equal(id)
+    )
+
+    val criteria = generalCriteria.and(existingId)
+    val result = serviceHub.vaultService.queryBy<IndyCredential>(criteria)
+
+    return result.states.firstOrNull()
 }
