+++ cordlib-contracts-states/src/main/kotlin/com/luxoft/blockchainlab/corda/hyperledger/indy/data/state/IndyRevocationRegistryDefinition.kt
@@ -4,8 +4,14 @@ import com.luxoft.blockchainlab.corda.hyperledger.indy.data.schema.RevocationReg
 import com.luxoft.blockchainlab.hyperledger.indy.models.CredentialDefinitionId
 import com.luxoft.blockchainlab.hyperledger.indy.models.RevocationRegistryDefinitionId
 import net.corda.core.contracts.LinearState
+import net.corda.core.contracts.StateAndRef
 import net.corda.core.contracts.UniqueIdentifier
+import net.corda.core.flows.FlowLogic
 import net.corda.core.identity.AbstractParty
+import net.corda.core.node.services.Vault
+import net.corda.core.node.services.queryBy
+import net.corda.core.node.services.vault.Builder.equal
+import net.corda.core.node.services.vault.QueryCriteria
 import net.corda.core.schemas.MappedSchema
 import net.corda.core.schemas.PersistentState
 import net.corda.core.schemas.QueryableState
@@ -48,4 +54,31 @@ data class IndyRevocationRegistryDefinition(
      * Returns new state
      */
     fun requestNewCredential() = copy(currentCredNumber = this.currentCredNumber + 1)
+}
+
+/**
+ * Gets revocation registry state from vault
+ */
+fun FlowLogic<Any>.getRevocationRegistryDefinitionById(id: RevocationRegistryDefinitionId): StateAndRef<IndyRevocationRegistryDefinition>? {
+    val generalCriteria = QueryCriteria.VaultQueryCriteria(Vault.StateStatus.UNCONSUMED)
+    val id = QueryCriteria.VaultCustomQueryCriteria(
+        RevocationRegistryDefinitionSchemaV1.PersistentRevocationRegistryDefinition::id.equal(id.toString())
+    )
+
+    val criteria = generalCriteria.and(id)
+    val result = serviceHub.vaultService.queryBy<IndyRevocationRegistryDefinition>(criteria)
+
+    return result.states.firstOrNull()
+}
+
+fun FlowLogic<Any>.getRevocationRegistryDefinitionByCredentialDefinitionId(id: CredentialDefinitionId): StateAndRef<IndyRevocationRegistryDefinition>? {
+    val generalCriteria = QueryCriteria.VaultQueryCriteria(Vault.StateStatus.UNCONSUMED)
+    val id = QueryCriteria.VaultCustomQueryCriteria(
+        RevocationRegistryDefinitionSchemaV1.PersistentRevocationRegistryDefinition::credentialDefinitionId.equal(id.toString())
+    )
+
+    val criteria = generalCriteria.and(id)
+    val result = serviceHub.vaultService.queryBy<IndyRevocationRegistryDefinition>(criteria)
+
+    return result.states.firstOrNull()
 }
