+++ cordlib-contracts-states/src/main/kotlin/com/luxoft/blockchainlab/corda/hyperledger/indy/data/state/IndySchema.kt
@@ -3,8 +3,14 @@ package com.luxoft.blockchainlab.corda.hyperledger.indy.data.state
 import com.luxoft.blockchainlab.corda.hyperledger.indy.data.schema.IndySchemaSchemaV1
 import com.luxoft.blockchainlab.hyperledger.indy.models.SchemaId
 import net.corda.core.contracts.LinearState
+import net.corda.core.contracts.StateAndRef
 import net.corda.core.contracts.UniqueIdentifier
+import net.corda.core.flows.FlowLogic
 import net.corda.core.identity.AbstractParty
+import net.corda.core.node.services.Vault
+import net.corda.core.node.services.queryBy
+import net.corda.core.node.services.vault.Builder.equal
+import net.corda.core.node.services.vault.QueryCriteria
 import net.corda.core.schemas.MappedSchema
 import net.corda.core.schemas.PersistentState
 import net.corda.core.schemas.QueryableState
@@ -31,4 +37,22 @@ class IndySchema(
     }
 
     override fun supportedSchemas() = listOf(IndySchemaSchemaV1)
+}
+
+/**
+ * Gets schema state from vault
+ */
+fun FlowLogic<Any>.getSchemaById(schemaId: SchemaId): StateAndRef<IndySchema>? {
+    val generalCriteria =
+        QueryCriteria.VaultQueryCriteria(Vault.StateStatus.UNCONSUMED)
+    val id = QueryCriteria.VaultCustomQueryCriteria(
+        IndySchemaSchemaV1.PersistentSchema::id.equal(
+            schemaId.toString()
+        )
+    )
+
+    val criteria = generalCriteria.and(id)
+    val result = serviceHub.vaultService.queryBy<IndySchema>(criteria)
+
+    return result.states.firstOrNull()
 }
