+++ cordlib/src/main/kotlin/com.luxoft.blockchainlab.corda.hyperledger.indy/flow/RevokeCredentialFlow.kt
@@ -2,6 +2,7 @@ package com.luxoft.blockchainlab.corda.hyperledger.indy.flow
 
 import co.paralleluniverse.fibers.Suspendable
 import com.luxoft.blockchainlab.corda.hyperledger.indy.contract.IndyCredentialContract
+import com.luxoft.blockchainlab.corda.hyperledger.indy.data.state.getIndyCredentialState
 import net.corda.core.contracts.Command
 import net.corda.core.flows.*
 import net.corda.core.transactions.TransactionBuilder
