+++ indy-utils/src/main/kotlin/com/luxoft/blockchainlab/hyperledger/indy/IndyFacade.kt
@@ -100,7 +100,7 @@ interface IndyFacade {
      * @param credentialRequest [CredentialRequestInfo] - credential request and all reliable data
      * @param offer [CredentialOffer]
      */
-    fun receiveCredential(
+    fun checkLedgerAndReceiveCredential(
         credentialInfo: CredentialInfo,
         credentialRequest: CredentialRequestInfo,
         offer: CredentialOffer
