+++ indy-utils/src/main/kotlin/com/luxoft/blockchainlab/hyperledger/indy/IndyUser.kt
@@ -103,7 +103,7 @@ class IndyUser(
         return credentialInfo
     }
 
-    override fun receiveCredential(
+    override fun checkLedgerAndReceiveCredential(
         credentialInfo: CredentialInfo,
         credentialRequest: CredentialRequestInfo,
         offer: CredentialOffer
